feat: add global styles
This commit adds the following global styles to the website:

Set the font size to 62.5% on the root element using the :root selector.
Set the height to 100% on the html and body elements.
Unset all default styles on p elements.
Set the font size to 1.6rem, display to flex, and other basic styles on the body element, along with a dark background color and white text color.
Inherit font family and set font size to 100% on form elements
Set width, padding, and margin to 0 and box-sizing to form elements